Refactor Resource class
