+++ packages/panels/src/Resources/Resource/Concerns/HasBreadcrumbs.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Filament\Resources\Resource\Concerns;
+
+trait HasBreadcrumbs
+{
+    protected static ?string $breadcrumb = null;
+
+    public static function getBreadcrumb(): string
+    {
+        return static::$breadcrumb ?? static::getTitleCasePluralModelLabel();
+    }
+}
